박대현 redis 연결명령어  서비스 아키텍쳐에서 분리
