+++ app.js
@@ -3,21 +3,6 @@ const app = express();
 const port = 5000;
 const http = require('http').createServer(app);
 const cors = require('cors')
-const redis = require('redis');
-const dotenv = require('dotenv');
-
-dotenv.config(); // env환경변수 파일 가져오기
-
-const redisClient = redis.createClient({
-   url: `redis://${process.env.REDIS_USERNAME}:${process.env.REDIS_PASSWORD}@${process.env.REDIS_HOST}:${process.env.REDIS_PORT}`,
-   legacyMode: true, // 반드시 설정 !!
-});
-redisClient.on('connect', () => {
-   console.info('Redis connected!');
-});
-redisClient.on('error', (err) => {
-   console.error('Redis Client Error', err);
-});
 
 app.use(express.static("assets"));
 
@@ -39,4 +24,3 @@ http.listen(port, () => {
     console.log(`${port}번 포트로 서버 실행`);
 });
 
-module.exports = redisClient
