+++ routes/SaveDatas.routes.js
@@ -4,10 +4,10 @@ const SaveDataController = require ('../controllers/SaveDatas.controller');
 
 const saveDataController = new SaveDataController()
 // /saveData/
-router.get('/summonerMatchlist', saveDataController.getSummoner) // 매치 리스트 저장
 router.get('/matchlist', saveDataController.getMatchList) // 매치 리스트 저장
-router.get('/matchData/:tier', saveDataController.saveMatchData) // 단일 매치 저장 
-router.put('/rating', saveDataController.saveRating)
+router.get('/summonerMatchlist', saveDataController.getSummoner) // 소환사 이름으로 단일 소환사 매치 리스트 저장
+router.get('/matchData/:tier', saveDataController.saveMatchData) // 매치데이터 업데이트
+router.put('/rating', saveDataController.saveRating) 
 
 
 module.exports = router;
