+++ services/Matches.service.js
@@ -3,25 +3,7 @@ const MatchesRepository = require('../repositories/Matches.repository');
 const championData = require('../dataInfo/champId.js');
 const coreItemList = require('../dataInfo/item.js')
 const API = require('../apiList');
-
-const redis = require('redis');
-const dotenv = require('dotenv');
-
-dotenv.config(); // env환경변수 파일 가져오기
-
-const redisClient = redis.createClient({
-   url: `redis://${process.env.REDIS_USERNAME}:${process.env.REDIS_PASSWORD}@${process.env.REDIS_HOST}:${process.env.REDIS_PORT}`,
-   legacyMode: true, // 반드시 설정 !!
-});
-redisClient.on('connect', () => {
-   console.info('Redis connected!');
-});
-redisClient.on('error', (err) => {
-   console.error('Redis Client Error', err);
-});
-redisClient.connect().then(); // redis v4 연결 (비동기)
-const redisCli = redisClient.v4;
-redisCli.flushAll() // redis db 전체 초기화
+const redisCli = require('../redis')
 
 class MatchesService {
     constructor() {
@@ -86,21 +68,11 @@ class MatchesService {
         const data = JSON.stringify(itemByEnemyResult)
 
         await redisCli.set(`${myChampionId}and${enemyChampionId}`,data)
-        await redisCli.expire(`${myChampionId}and${enemyChampionId}`,10) // 만료 시간 설정 int 는 sec 
+        await redisCli.expire(`${myChampionId}and${enemyChampionId}`,10) // 만료 시간 설정 int 는 sec 원래대로라면 전적갱신을 눌렀을때 캐시데이터도 갱신을 해야함
 
         return itemByEnemyResult
     }
     };
-
-    getItem = async(itemId) => {
-        const item = await this.matchesRepository.getItemById(itemId)
-
-        item.map((data) => {
-            
-        })
-
-        return item
-    }
 
     getSummoner = async (summonerName) => {
